content/media/video/test/test_constants.html

<!DOCTYPE HTML>
<html>
<!--
  Adapted from:
  http://simon.html5.org/test/html/dom/interfaces/HTMLElement/HTMLMediaElement/const-unsigned-short/001.htm
-->
<head>
  <title>Media test: constants</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video><source></video><audio><source></audio>
<pre id="test">
<script class="testbody" type="text/javascript">
is(HTMLElement.EMPTY, undefined);
is(HTMLElement.LOADING, undefined);
is(HTMLElement.LOADED_METADATA, undefined);
is(HTMLElement.LOADED_FIRST_FRAME, undefined);
is(HTMLElement.LOADED, undefined);
is(HTMLElement.DATA_UNAVAILABLE, undefined);
is(HTMLElement.CAN_SHOW_CURRENT_FRAME, undefined);
is(HTMLElement.CAN_PLAY, undefined);
is(HTMLElement.CAN_PLAY_THROUGH, undefined);
is(HTMLMediaElement.EMPTY, 0);
is(HTMLMediaElement.LOADING, 1);
is(HTMLMediaElement.LOADED_METADATA, 2);
is(HTMLMediaElement.LOADED_FIRST_FRAME, 3);
is(HTMLMediaElement.LOADED, 4);
is(HTMLMediaElement.DATA_UNAVAILABLE, 0);
is(HTMLMediaElement.CAN_SHOW_CURRENT_FRAME, 1);
is(HTMLMediaElement.CAN_PLAY, 2);
is(HTMLMediaElement.CAN_PLAY_THROUGH, 3);
is(HTMLVideoElement.EMPTY, undefined);
is(HTMLVideoElement.LOADING, undefined);
is(HTMLVideoElement.LOADED_METADATA, undefined);
is(HTMLVideoElement.LOADED_FIRST_FRAME, undefined);
is(HTMLVideoElement.LOADED, undefined);
is(HTMLVideoElement.DATA_UNAVAILABLE, undefined);
is(HTMLVideoElement.CAN_SHOW_CURRENT_FRAME, undefined);
is(HTMLVideoElement.CAN_PLAY, undefined);
is(HTMLVideoElement.CAN_PLAY_THROUGH, undefined);
is(HTMLAudioElement.EMPTY, undefined);
is(HTMLAudioElement.LOADING, undefined);
is(HTMLAudioElement.LOADED_METADATA, undefined);
is(HTMLAudioElement.LOADED_FIRST_FRAME, undefined);
is(HTMLAudioElement.LOADED, undefined);
is(HTMLAudioElement.DATA_UNAVAILABLE, undefined);
is(HTMLAudioElement.CAN_SHOW_CURRENT_FRAME, undefined);
is(HTMLAudioElement.CAN_PLAY, undefined);
is(HTMLAudioElement.CAN_PLAY_THROUGH, undefined);
is(HTMLSourceElement.EMPTY, undefined);
is(HTMLSourceElement.LOADING, undefined);
is(HTMLSourceElement.LOADED_METADATA, undefined);
is(HTMLSourceElement.LOADED_FIRST_FRAME, undefined);
is(HTMLSourceElement.LOADED, undefined);
is(HTMLSourceElement.DATA_UNAVAILABLE, undefined);
is(HTMLSourceElement.CAN_SHOW_CURRENT_FRAME, undefined);
is(HTMLSourceElement.CAN_PLAY, undefined);
is(HTMLSourceElement.CAN_PLAY_THROUGH, undefined);
is(document.body.EMPTY, undefined);
is(document.body.LOADING, undefined);
is(document.body.LOADED_METADATA, undefined);
is(document.body.LOADED_FIRST_FRAME, undefined);
is(document.body.LOADED, undefined);             
is(document.body.DATA_UNAVAILABLE, undefined);
is(document.body.CAN_SHOW_CURRENT_FRAME, undefined);
is(document.body.CAN_PLAY, undefined);
is(document.body.CAN_PLAY_THROUGH, undefined);
is(document.getElementsByTagName("video")[0].EMPTY, 0);
is(document.getElementsByTagName("video")[0].LOADING, 1);
is(document.getElementsByTagName("video")[0].LOADED_METADATA, 2);
is(document.getElementsByTagName("video")[0].LOADED_FIRST_FRAME, 3);
is(document.getElementsByTagName("video")[0].LOADED, 4);
is(document.getElementsByTagName("video")[0].DATA_UNAVAILABLE, 0);
is(document.getElementsByTagName("video")[0].CAN_SHOW_CURRENT_FRAME, 1);
is(document.getElementsByTagName("video")[0].CAN_PLAY, 2);
is(document.getElementsByTagName("video")[0].CAN_PLAY_THROUGH, 3);
is(document.getElementsByTagName("audio")[0].EMPTY, 0);
is(document.getElementsByTagName("audio")[0].LOADING, 1);
is(document.getElementsByTagName("audio")[0].LOADED_METADATA, 2);
is(document.getElementsByTagName("audio")[0].LOADED_FIRST_FRAME, 3);
is(document.getElementsByTagName("audio")[0].LOADED, 4);
is(document.getElementsByTagName("audio")[0].DATA_UNAVAILABLE, 0);
is(document.getElementsByTagName("audio")[0].CAN_SHOW_CURRENT_FRAME, 1);
is(document.getElementsByTagName("audio")[0].CAN_PLAY, 2);
is(document.getElementsByTagName("audio")[0].CAN_PLAY_THROUGH, 3);
is(document.getElementsByTagName("source")[0].EMPTY, undefined);
is(document.getElementsByTagName("source")[0].LOADING, undefined);
is(document.getElementsByTagName("source")[0].LOADED_METADATA, undefined);
is(document.getElementsByTagName("source")[0].LOADED_FIRST_FRAME, undefined);
is(document.getElementsByTagName("source")[0].LOADED, undefined);
is(document.getElementsByTagName("source")[0].DATA_UNAVAILABLE, undefined);
is(document.getElementsByTagName("source")[0].CAN_SHOW_CURRENT_FRAME, undefined);
is(document.getElementsByTagName("source")[0].CAN_PLAY, undefined);
is(document.getElementsByTagName("source")[0].CAN_PLAY_THROUGH, undefined);
is(document.getElementsByTagName("source")[1].EMPTY, undefined);
is(document.getElementsByTagName("source")[1].LOADING, undefined);
is(document.getElementsByTagName("source")[1].LOADED_METADATA, undefined);
is(document.getElementsByTagName("source")[1].LOADED_FIRST_FRAME, undefined);
is(document.getElementsByTagName("source")[1].LOADED, undefined);
is(document.getElementsByTagName("source")[1].DATA_UNAVAILABLE, undefined);
is(document.getElementsByTagName("source")[1].CAN_SHOW_CURRENT_FRAME, undefined);
is(document.getElementsByTagName("source")[1].CAN_PLAY, undefined);
is(document.getElementsByTagName("source")[1].CAN_PLAY_THROUGH, undefined);
is(HTMLElement.prototype.EMPTY, undefined);
is(HTMLElement.prototype.LOADING, undefined);
is(HTMLElement.prototype.LOADED_METADATA, undefined);
is(HTMLElement.prototype.LOADED_FIRST_FRAME, undefined);
is(HTMLElement.prototype.LOADED, undefined);
is(HTMLElement.prototype.DATA_UNAVAILABLE, undefined);
is(HTMLElement.prototype.CAN_SHOW_CURRENT_FRAME, undefined);
is(HTMLElement.prototype.CAN_PLAY, undefined);
is(HTMLElement.prototype.CAN_PLAY_THROUGH, undefined);
is(HTMLVideoElement.prototype.EMPTY, 0);
is(HTMLVideoElement.prototype.LOADING, 1);
is(HTMLVideoElement.prototype.LOADED_METADATA, 2);
is(HTMLVideoElement.prototype.LOADED_FIRST_FRAME, 3);
is(HTMLVideoElement.prototype.LOADED, 4);
is(HTMLVideoElement.prototype.DATA_UNAVAILABLE, 0);
is(HTMLVideoElement.prototype.CAN_SHOW_CURRENT_FRAME, 1);
is(HTMLVideoElement.prototype.CAN_PLAY, 2);
is(HTMLVideoElement.prototype.CAN_PLAY_THROUGH, 3);
is(HTMLAudioElement.prototype.EMPTY, 0);
is(HTMLAudioElement.prototype.LOADING, 1);
is(HTMLAudioElement.prototype.LOADED_METADATA, 2);
is(HTMLAudioElement.prototype.LOADED_FIRST_FRAME, 3);
is(HTMLAudioElement.prototype.LOADED, 4);
is(HTMLAudioElement.prototype.DATA_UNAVAILABLE, 0);
is(HTMLAudioElement.prototype.CAN_SHOW_CURRENT_FRAME, 1);
is(HTMLAudioElement.prototype.CAN_PLAY, 2);
is(HTMLAudioElement.prototype.CAN_PLAY_THROUGH, 3);
is(HTMLSourceElement.prototype.EMPTY, undefined);
is(HTMLSourceElement.prototype.LOADING, undefined);
is(HTMLSourceElement.prototype.LOADED_METADATA, undefined);
is(HTMLSourceElement.prototype.LOADED_FIRST_FRAME, undefined);
is(HTMLSourceElement.prototype.LOADED, undefined);
is(HTMLSourceElement.prototype.DATA_UNAVAILABLE, undefined);
is(HTMLSourceElement.prototype.CAN_SHOW_CURRENT_FRAME, undefined);
is(HTMLSourceElement.prototype.CAN_PLAY, undefined);
is(HTMLSourceElement.prototype.CAN_PLAY_THROUGH, undefined);
</script>
</pre>
</body>
</html>
